Move a bunch more filter code into @weco/catalogue
