+++ catalogue/webapp/pages/search/works.tsx
@@ -27,7 +27,7 @@ import { getServerData } from '@weco/common/server-data';
 import { NextPageWithLayout } from '@weco/common/views/pages/_app';
 import { Pageview } from '@weco/common/services/conversion/track';
 import { getWorks } from '@weco/catalogue/services/catalogue/works';
-import { worksFilters } from '@weco/common/services/catalogue/filters';
+import { worksFilters } from '@weco/catalogue/services/catalogue/filters';
 import convertUrlToString from '@weco/common/utils/convert-url-to-string';
 import { hasFilters, linkResolver } from '@weco/common/utils/search';
 import { AppErrorProps, appError } from '@weco/common/services/app';
