+++ catalogue/webapp/services/catalogue/fixtures/images-aggregations.ts
@@ -1,4 +1,4 @@
-import { CatalogueResultsList, Image } from '../../../model/catalogue';
+import { CatalogueResultsList, Image } from '@weco/common/model/catalogue';
 
 const aggregations: CatalogueResultsList<Image> = {
   type: 'ResultList',
@@ -71,6 +71,7 @@ const aggregations: CatalogueResultsList<Image> = {
     },
     type: 'Aggregations',
   },
+  _requestUrl: 'https://api.wellcomecollection.org/catalogue/v2/images',
 };
 
 export default aggregations;
